+++ test_proj/SimpleLogic/src/main/java/com/sam/quest/dao/UserDAO.java
@@ -5,6 +5,6 @@ import  com.sam.quest.entity.Users;
 public interface UserDAO {
     public boolean insertUser(Users user);
     public boolean deleteUser(Users user);
-    public Users findUser(int userId);
+    public Users findUser(long userId);
     public boolean updateUser(Users user);
 }
